delete the loading more on scrolling: now loading 100 items by default
