+++ app/src/main/java/net/squanchy/tweets/TweetsPageView.java
@@ -2,7 +2,6 @@ package net.squanchy.tweets;
 
 import android.content.Context;
 import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.widget.LinearLayout;
@@ -10,7 +9,6 @@ import android.widget.TextView;
 
 import net.squanchy.R;
 import net.squanchy.tweets.service.TwitterRepository;
-import net.squanchy.tweets.view.ScrollListener;
 import net.squanchy.tweets.view.TweetsAdapter;
 
 import io.reactivex.disposables.Disposable;
@@ -22,7 +20,6 @@ public class TweetsPageView extends LinearLayout {
     private RecyclerView tweetsList;
     private TweetsAdapter tweetsAdapter;
     private SwipeRefreshLayout swipeLayout;
-    private TweetScrollListener scrollListener;
     private Disposable disposable;
     private boolean refreshingData;
 
@@ -70,7 +67,6 @@ public class TweetsPageView extends LinearLayout {
     @Override
     protected void onDetachedFromWindow() {
         super.onDetachedFromWindow();
-        scrollListener.destroy();
         if (disposable != null && !disposable.isDisposed()) {
             disposable.dispose();
         }
@@ -87,9 +83,6 @@ public class TweetsPageView extends LinearLayout {
 
         emptyView.setText(context.getString(R.string.no_tweets_for_query, query));
 
-        LinearLayoutManager layoutManager = (LinearLayoutManager) tweetsList.getLayoutManager();
-        scrollListener = new TweetScrollListener(layoutManager);
-        tweetsList.addOnScrollListener(scrollListener);
         refreshTimeline();
     }
 
@@ -118,31 +111,5 @@ public class TweetsPageView extends LinearLayout {
     private void onError(Throwable throwable) {
         Timber.e(throwable);
         onRefreshFinished();
-    }
-
-    private class TweetScrollListener extends ScrollListener {
-
-        private Disposable disposable;
-
-        TweetScrollListener(LinearLayoutManager layoutManager) {
-            super(layoutManager);
-        }
-
-        @Override
-        protected void loadMore() {
-            if (refreshingData) {
-                return;
-            }
-            disposable = tweetsAdapter.previous()
-                    .doOnSubscribe(v -> loading = true)
-                    .doOnTerminate(() -> loading = false)
-                    .subscribe(search -> onRefreshFinished(), TweetsPageView.this::onError);
-        }
-
-        void destroy() {
-            if (disposable != null && !disposable.isDisposed()) {
-                disposable.dispose();
-            }
-        }
     }
 }
