--- app/src/main/java/net/squanchy/tweets/view/LoadingViewHolder.java
@@ -1,11 +1,0 @@
-package net.squanchy.tweets.view;
-
-import android.support.v7.widget.RecyclerView;
-import android.view.View;
-
-class LoadingViewHolder extends RecyclerView.ViewHolder {
-
-    LoadingViewHolder(View itemView) {
-        super(itemView);
-    }
-}
