--- app/src/main/res/layout/item_tweet_loading.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ProgressBar xmlns:android="http://schemas.android.com/apk/res/android"
-  style="@style/Base.Widget.AppCompat.ProgressBar"
-  android:layout_width="match_parent"
-  android:layout_height="wrap_content" />
